mysql导入优化

+++ plugins/mysql-yum/index.py
@@ -670,8 +670,7 @@ def importDbExternal():
         return mw.returnJson(False, '文件突然消失?')
 
     exts = ['sql', 'gz', 'zip']
-    tmp = file.split('.')
-    ext = tmp[len(tmp) - 1]
+    ext = mw.getFileSuffix(file)
     if ext not in exts:
         return mw.returnJson(False, '导入数据库格式不对!')
 
@@ -700,6 +699,9 @@ def importDbExternal():
         mw.execShell(cmd)
         import_sql = import_dir + tmpFile
 
+    if file.find(".sql") > -1 and file.find(".sql.gz") == -1:
+        import_sql = import_dir + file
+
     if import_sql == "":
         return mw.returnJson(False, '未找SQL文件')
 
@@ -712,7 +714,8 @@ def importDbExternal():
 
     # print(mysql_cmd)
     os.system(mysql_cmd)
-    os.remove(import_sql)
+    if ext != 'sql':
+        os.remove(import_sql)
 
     return mw.returnJson(True, 'ok')
